refacotr: add RewritePartitions Interface
