+++ core/src/main/java/com/netease/arctic/op/PartitionTransactionOperation.java
@@ -22,6 +22,7 @@ import com.netease.arctic.table.BaseTable;
 import com.netease.arctic.table.KeyedTable;
 import com.netease.arctic.table.TableProperties;
 import com.netease.arctic.utils.TablePropertyUtil;
+import org.apache.iceberg.PendingUpdate;
 import org.apache.iceberg.Transaction;
 import org.apache.iceberg.UpdateProperties;
 import org.apache.iceberg.util.StructLikeMap;
@@ -30,9 +31,11 @@ import org.apache.iceberg.util.StructLikeMap;
  * Abstract transaction operation on {@link BaseTable} which will change
  * max transaction id map
  */
-public abstract class PartitionTransactionOperation {
+public abstract class PartitionTransactionOperation implements PendingUpdate<StructLikeMap<Long>> {
 
   KeyedTable keyedTable;
+  private Transaction tx;
+
 
   public PartitionTransactionOperation(KeyedTable baseTable) {
     this.keyedTable = baseTable;
@@ -47,10 +50,16 @@ public abstract class PartitionTransactionOperation {
    */
   protected abstract StructLikeMap<Long> apply(Transaction transaction, StructLikeMap<Long> partitionMaxTxId);
 
-  public void commit() {
-    Transaction tx = keyedTable.baseTable().newTransaction();
+  @Override
+  public StructLikeMap<Long> apply() {
+    return apply(tx, keyedTable.partitionMaxTransactionId());
+  }
 
-    StructLikeMap<Long> partitionMaxTxId = apply(tx, keyedTable.partitionMaxTransactionId());
+
+  public void commit() {
+    this.tx = keyedTable.baseTable().newTransaction();
+
+    StructLikeMap<Long> partitionMaxTxId = apply();
 
     String propertyValue = TablePropertyUtil.encodePartitionMaxTxId(keyedTable.spec(), partitionMaxTxId);
     UpdateProperties updateProperties = tx.updateProperties();
